# internal/swagger/combined.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EXBanka 3 Gateway API",
    "version": "2026-03-25",
    "description": "Current gateway-facing route surface for the EXBanka 3 stack. Business endpoints are served through nginx on /api/v1 and routed to the underlying microservices."
  },
  "servers": [
    {
      "url": "http://localhost",
      "description": "Local nginx gateway"
    }
  ],
...
  "paths": {
...
    "/api/v1/clients/{id}": {
      "get": {
        "tags": ["Clients"],
        "summary": "Get client details",
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": { "description": "Client details returned" }
        }
      },
      "put": {
        "tags": ["Clients"],
        "summary": "Update client",
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": { "description": "Client updated" }
        }
      }
    },
...
    "/api/v1/accounts/client/{clientId}": {
      "get": {
        "tags": ["Accounts"],
        "summary": "List client accounts",
        "parameters": [{ "$ref": "#/components/parameters/ClientIdPath" }],
        "responses": {
          "200": { "description": "Client accounts returned" }
        }
      }
    },
    "/api/v1/accounts/{id}": {
      "get": {
        "tags": ["Accounts"],
        "summary": "Get account details",
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": { "description": "Account returned" }
        }
      }
    },
...
    "/api/v1/cards/account/{accountId}": {
      "get": {
        "tags": ["Cards"],
        "summary": "List cards by account",
        "parameters": [{ "$ref": "#/components/parameters/AccountIdPath" }],
        "responses": {
          "200": { "description": "Account cards returned" }
        }
      }
    },
    "/api/v1/cards/client/{clientId}": {
      "get": {
        "tags": ["Cards"],
        "summary": "List cards visible to a client",
        "parameters": [{ "$ref": "#/components/parameters/ClientIdPath" }],
        "responses": {
          "200": { "description": "Client cards returned" }
        }
      }
    },
...
    "/api/v1/transfers/client/{clientId}": {
      "get": {
        "tags": ["Transfers"],
        "summary": "List transfers for a client",
        "parameters": [{ "$ref": "#/components/parameters/ClientIdPath" }],
        "responses": {
          "200": { "description": "Client transfers returned" }
        }
      }
    },
    "/api/v1/transfers/account/{accountId}": {
      "get": {
        "tags": ["Transfers"],
        "summary": "List transfers for an account",
        "parameters": [{ "$ref": "#/components/parameters/AccountIdPath" }],
        "responses": {
          "200": { "description": "Account transfers returned" }
        }
      }
    },
...
    "/api/v1/payments/client/{clientId}": {
      "get": {
        "tags": ["Payments"],
        "summary": "List payments for a client",
        "parameters": [{ "$ref": "#/components/parameters/ClientIdPath" }],
        "responses": {
          "200": { "description": "Client payments returned" }
        }
      }
    },
    "/api/v1/payments/account/{accountId}": {
      "get": {
        "tags": ["Payments"],
        "summary": "List payments for an account",
        "parameters": [{ "$ref": "#/components/parameters/AccountIdPath" }],
        "responses": {
          "200": { "description": "Account payments returned" }
        }
      }
    },
...
    "/api/v1/loans/client/{clientId}": {
      "get": {
        "tags": ["Loans"],
        "summary": "List loans for a client",
        "parameters": [{ "$ref": "#/components/parameters/ClientIdPath" }],
        "responses": {
          "200": { "description": "Client loans returned" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      },
      "ClientIdPath": {
        "name": "clientId",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      },
      "AccountIdPath": {
        "name": "accountId",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      }
    }
  }
}